This is a code created to play the game of guessing the number betwen user and computer.
At the game beginning the user is instructed to think a number between 0 and 100.

And then after delay the computer starts guessing the number.
the user has to input "Y" for a correct guess or "H" for a number guessed greater then user's number.
or "L" for a number guessed lesser then user's number.

the game continues till the computer guesses the correct number.

At the end When the computer guesses the correct number it Displays "How many guesses it took to guess the number".